scripts/requirements.txt
PyYAML
python-dateutil